Dernier commit - Mise en place de la modale de supression

+++ src/index.js
@@ -1,10 +1,11 @@
 import "./assets/styles/styles.scss";
 import "./index.scss";
+import "./assets/javascripts/topbar.js";
+import { openModal } from "./assets/javascripts/modal";
 
 const articleContainerElement = document.querySelector(".articles-container");
 const categoriesContainerElement = document.querySelector(".categories");
 const selectElement = document.querySelector("select");
-
 let filter;
 let articles;
 let sortBy = "desc";
@@ -58,26 +59,29 @@ const createArticles = () => {
     button.addEventListener("click", (event) => {
       const target = event.target;
       const articleId = target.dataset.id;
-      // Lien pour l'éditeur en ligne
-      window.location.assign(`src/form/form.html?id=${articleId}`);
+      location.assign(`/form/form.html?id=${articleId}`);
     });
   });
   deleteButtons.forEach((button) => {
     button.addEventListener("click", async (event) => {
-      try {
-        const target = event.target;
-        const articleId = target.dataset.id;
-        const response = await fetch(
-          `https://restapi.fr/api/article/${articleId}`,
-          {
-            method: "DELETE",
-          }
-        );
-        const body = await response.json();
-        console.log(body);
-        fetchArticle();
-      } catch (e) {
-        console.log("e : ", e);
+      const result = await openModal(
+        "Etes vous sur de vouloir supprimer votre article ?"
+      );
+      if (result === true) {
+        try {
+          const target = event.target;
+          const articleId = target.dataset.id;
+          const response = await fetch(
+            `https://restapi.fr/api/article/${articleId}`,
+            {
+              method: "DELETE",
+            }
+          );
+          const body = await response.json();
+          fetchArticle();
+        } catch (e) {
+          console.log("e : ", e);
+        }
       }
     });
   });
@@ -94,14 +98,15 @@ const displayMenuCategories = (categoriesArr) => {
       if (filter === categoryElem[0]) {
         filter = null;
         li.classList.remove("active");
+        createArticles();
       } else {
         filter = categoryElem[0];
         liElements.forEach((li) => {
           li.classList.remove("active");
         });
         li.classList.add("active");
+        createArticles();
       }
-      createArticles();
     });
     return li;
   });
@@ -134,6 +139,9 @@ const fetchArticle = async () => {
       `https://restapi.fr/api/article?sort=createdAt:${sortBy}`
     );
     articles = await response.json();
+    if (!Array.isArray(articles)) {
+      articles = [articles];
+    }
     createArticles();
     createMenuCategories();
   } catch (e) {
